api/requirements.txt
attrs==19.3.0
boto3==1.10.30
botocore==1.13.30
Click==7.0
docutils==0.15.2
Flask==1.1.1
importlib-metadata==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
more-itertools==8.0.0
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.13.0
urllib3==1.25.7
waitress==1.3.1
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
